Don't use filename as invoice number
