+++ invoicez/cli/new.py
@@ -2,7 +2,9 @@ from datetime import datetime, timedelta
 from logging import getLogger
 from pathlib import Path
 
-from invoicez.cli import command, dir_path_option
+from yaml import safe_dump as yaml_safe_dump, safe_load as yaml_safe_load
+
+from invoicez.cli import command, dir_path_option, path_argument
 from invoicez.paths import Paths
 
 
@@ -10,9 +12,10 @@ _logger = getLogger(__name__)
 
 
 @command
+@path_argument
 @dir_path_option
-def new(dir_path: str) -> None:
-    """Create a new invoice."""
+def new(path: str, dir_path: str) -> None:
+    """Create a new invoice based on an existing one."""
     paths = Paths(Path(dir_path))
     now = datetime.now()
     prefix = now.strftime("%Y-%m")
@@ -20,10 +23,12 @@ def new(dir_path: str) -> None:
     limit_date = (now + timedelta(days=31)).strftime("%d/%m/%Y")
     n = len(list(paths.git_dir.glob(f"*/{prefix}-*.yml"))) + 1
     output_path = paths.working_dir / f"{prefix}-{n:03}.yml"
-    content = paths.template_invoice.read_text(encoding="utf8")
-    content = content.replace("{DATE}", date)
-    content = content.replace("{LIMIT_DATE}", limit_date)
+    with (paths.working_dir / path).open(encoding="utf8") as fh:
+        content = yaml_safe_load(fh)
+    content["date"] = date
+    content["limit_date"] = limit_date
     _logger.info(
-        f"Copying invoice template to {output_path.relative_to(paths.working_dir)}"
+        f"Creating new invoice in {output_path.relative_to(paths.working_dir)}"
     )
-    output_path.write_text(content, encoding="utf8")
+    with output_path.open("w", encoding="utf8") as fh:
+        yaml_safe_dump(content, fh, allow_unicode=True)
