+++ invoicez/cli/run.py
@@ -1,23 +1,9 @@
 from pathlib import Path
 from typing import Any, List
 
-from click import argument, Path as ClickPath
-
-from invoicez.cli import command, dir_path_option, option
+from invoicez.cli import command, dir_path_option, option, path_argument
 from invoicez.paths import Paths
 from invoicez.runner import run as runner_run
-
-
-def _autocomplete_path(ctx: Any, args: List[str], incomplete: str) -> List[str]:
-    try:
-        paths = Paths(Path("."))
-        return [
-            str(t.relative_to(paths.working_dir))
-            for t in paths.working_dir.glob("*.yml")
-            if t.name != "company-config.yml"
-        ]
-    except Exception:
-        return []
 
 
 def _autocomplete_template(ctx: Any, args: List[str], incomplete: str) -> List[str]:
@@ -32,11 +18,7 @@ def _autocomplete_template(ctx: Any, args: List[str], incomplete: str) -> List[s
 
 
 @command
-@argument(
-    "path",
-    type=ClickPath(exists=True, dir_okay=False, readable=True),
-    autocompletion=_autocomplete_path,  # type: ignore
-)
+@path_argument
 @option(
     "--template",
     type=str,
