+++ pyproject.toml
@@ -6,7 +6,7 @@ homepage = "https://github.com/m09/invoicez"
 license = "Apache-2.0"
 name = "invoicez"
 readme = "README.md"
-version = "1.0.1"
+version = "2.0.0"
 
 [tool.poetry.dependencies]
 GitPython = "^3.1.0"
